requirements.txt
Flask==3.0.2
crossplane==0.5.8
python-dotenv==1.0.1

